Fixed Shared VPC / Add NetworkClasses(?)
