+++ sp/src/game/shared/modular_weapons/basemodularweapon.cpp
@@ -6,12 +6,35 @@
 //  Author: Nbc66
 //	Last Modified: 2025-2-24
 //------------------------------------------------------------------------------------------//
+
 #include "cbase.h"
 
 #include "basemodularweapon.h"
 #include "baseattachment.h"
 
 #include "tier0/memdbgon.h"
+
+IMPLEMENT_NETWORKCLASS_ALIASED(BaseModularWeapon, DT_BaseModularWeapon)
+
+BEGIN_NETWORK_TABLE(CBaseModularWeapon, DT_BaseModularWeapon)
+#ifdef GAME_DLL
+SendPropExclude("DT_AnimTimeMustBeFirst", "m_flAnimTime"),
+SendPropExclude("DT_BaseAnimating", "m_nSequence"),
+#endif
+END_NETWORK_TABLE()
+
+#ifdef CLIENT_DLL
+BEGIN_PREDICTION_DATA(CBaseModularWeapon)
+DEFINE_PRED_FIELD(m_flTimeWeaponIdle, FIELD_FLOAT, FTYPEDESC_OVERRIDE | FTYPEDESC_NOERRORCHECK),
+END_PREDICTION_DATA()
+#endif
+
+//LINK_ENTITY_TO_CLASS(weapon_base, CBaseModularWeapon);
+
+#ifdef GAME_DLL
+BEGIN_DATADESC(CBaseModularWeapon)
+END_DATADESC()
+#endif
 
 CBaseModularWeapon::CBaseModularWeapon()
 {
